Complete revamp to PHP version 8
- Refactor of CommandsBase.php
- Addressed a potential :bug: in DatabaseCommands.php when a table didn't have a PK
- Also in DatabaseCommands.php Added more output for `dbDetils()`
- Code clean up in ModelValidatorBase.php
- Removal of UserReplies.php since all the methods were moved to CommandsBase.php
