+++ app/Controllers/ModelValidatorBase.php
@@ -19,9 +19,7 @@ abstract class ModelValidatorBase
 {
     #[ApplyOverride('Override this property with the class name of the model to validate.')]
     protected string $modelClass;
-
     private static ?array $modelColumnAttributes = null;
-
     private static ?ReflectionClass $modelReflectionClass =  null;
 
     /**
